Add ViRulerHelperView.
This view abstracts some things that both ViLineNumberView and
ViFoldMarginView will need. Also move ViLineNumberView to ViRulerHelperView.
